Mobt 157 nbhood refactor consolidate unit tests part1 (#1665)
* Make options for neighbourhood sum more consistent.

* Consolidate percentile from neighbourhood unit tests

* Made weighted mode only a valid option if you are using circular kernel

* Minor changes following review comments

* Remove print statement
